requirements.txt
beautifulsoup4
requests
pillow
imagehash
